feat(suspense): use Next.js loading.js to wrap search page with suspense component
